+++ thelibrary-core/src/test/java/com/benfante/javacourse/thelibrary/core/model/BookTest.java
@@ -16,7 +16,7 @@ public class BookTest {
 		Author author = new Author(1, "The", "single author");
 		Publisher publisher = new Publisher(1, "A publisher");
 		float price = 1.23f;
-		Book book = new Book(id, title, author, publisher, price);
+		Book book = new Book(id, title, author, publisher, BigDecimal.valueOf(price));
 		assertEquals(id, book.getId());
 		assertEquals(title, book.getTitle());
 		assertEquals(author, book.getAuthor()[0]);
@@ -30,7 +30,7 @@ public class BookTest {
 		Author author = new Author(1, "The", "single author");
 		Publisher publisher = new Publisher(1, "A publisher");
 		float price = 1.23f;
-		Book book = new Book(id, title, author, publisher, price);
+		Book book = new Book(id, title, author, publisher, BigDecimal.valueOf(price));
 		long newId = 2;
 		String newTitle = "Another title";
 		Author newAuthor = new Author(2, "Another", "author");
@@ -72,11 +72,11 @@ public class BookTest {
 		float price = 1.23f;
 		String title = "A title";
 		
-		Book book = new Book(id, title, author, publisher, price);
+		Book book = new Book(id, title, author, publisher, BigDecimal.valueOf(price));
 		Author auth2 = new Author(1,"Pippo","Baudo");
 		assertFalse(book.hasAuthor(auth2));
 		//System.out.println(book.getPrint());
-		Book book2 = new Book(id, title, author, publisher, price);
+		Book book2 = new Book(id, title, author, publisher, BigDecimal.valueOf(price));
 		assertEquals(book,book2);
 		
 	}
@@ -89,8 +89,8 @@ public class BookTest {
 		Author author = new Author(1, "The", "single author");
 		Publisher publisher = new Publisher(1, "A publisher");
 		float price = 1.23f;
-		Book book = new Book(id, title, author, publisher, price);
-		Book book2 = new Book(id, title, author, publisher, price);
+		Book book = new Book(id, title, author, publisher, BigDecimal.valueOf(price));
+		Book book2 = new Book(id, title, author, publisher, BigDecimal.valueOf(price));
 		assertEquals(book,book2);
 		assertEquals(book.hashCode(),book2.hashCode());
 		book2.setId(3);
@@ -104,7 +104,7 @@ public class BookTest {
 		Author author = new Author(1, "The", "single author");
 		Publisher publisher = new Publisher(1, "A publisher");
 		float price = 1.23f;
-		Book book = new Book(id, title, author, publisher, price);
+		Book book = new Book(id, title, author, publisher, BigDecimal.valueOf(price));
 		book.addCategory(BookCategory.ARTS_AND_PHOTOGRAPHY);
 		assertTrue(Arrays.equals(book.getCategories(),new BookCategory[] {BookCategory.ARTS_AND_PHOTOGRAPHY}));
 		BookCategory[] cat = new BookCategory[] {BookCategory.HISTORY,BookCategory.LITERATURE_AND_FICTION,BookCategory.OTHER};
